Use the properly rendered room names and user names now available in the lib.
Apart from obvious changes, we had to tweak Message::Message() a bit. A Message object is in the room context by definition, so we are supposed to use Room::roomMemberName() instead of User::displayName() now. However, the respective Event object doesn't usually have the room attribute (frankly speaking, it's a shame - we should propagate this attribute from the upper-level event to contained events in the lib code). With this commit, a pointer to the room is passed into Message::Message().