Implemented room names and user names according to the CS spec.
1. User names are now disambiguated in the context of a specific room.
To avoid quadratic complexity, that required converting a flat list of
users in Room to a hashmap of members' displayNames to user object
pointers and changing all access to that list.
2. Room names are generated (according to the CS spec) out of
displayname, canonicalAlias, and member names (current, and then left).
Because of relying on member names, a room display name is now revised
every time the userlist of a room changes.