Mark a room unread only if the new events are messages
This changes three methods in QuaternionRoom.

In setShown we now reset the readmarker whenever it is not at the
bottom. And still delete the unread flag if it is set.

In doAddNewMessageEvents we now set the room unread only if one of the new
event really is a message.

In processEphemeralEvents we now only act if the event is a receipt.
If it is and there is no real message between the end and our own
readmarker we delete the unread flag.
If we can't find the readmarker we don't delete the unread flag, since
we can't tell if the server has unread messages.