Move the shown messages tracking logic from MessageEventModel to ChatRoomWidget
Along the way, changed the code so that timeline iterators are _really_ not stored; otherwise there can be a race condition when a maybe-read iterator is saved and then, e.g. older messages get loaded, invalidating timelineEdge(). The new code is more readable, as a free bonus ( readMarkerOnScreen is much clearer than maybeReadMessage != m_currentRoom->timelineEdge() ).