Improve chat edit widget
Use KChatEdit rather than a custom QLineEdit. KChatEdit is designed as
general-purpose input widget for chat applications, not Matrix-specific.
It has built-in support for history. Tab completion is achieved by
installing a simple event filter.

Rich-text is disabled, for now. We need support for it in the backend
library first.