Use QTextDocument* in the chatedit api
Previously the chat edit widget provided only QString objects, containing
either plain text or html. Now the api is much more flexbile, the caller
can choose what to do with the QTextDocument.

Another change in behavior is that pressing Return no longer clears the
input and puts it in the history. A returnPressed() signal notifies when
Return is pressed, and a new saveInput() method can be used to
explicitly accept an input and store it in the history.

Fixes #142