Fix tab completion
The [] QString operator has many overloads, one of them returning a
QCharRef. This particular overload might point to an invalidated
reference, because selectedText() is a temporary QString. This happens
at least on Linux platforms, where selectedText()[0] would return a null
QCharRef. Replacing the [] operator with .at(0) ensures that we get a
const QChar, which is what we want.