Updated to the latest lib; rewritten ImageProvider using a lambda function
Thanks to lambda captures, there's no more need of a hashmap to find the request data in the response handler. Moreover, the response handler is lockless now.
The new lib uses Qt logging categories, README.md has got a note on that.
Bonus: cleanup in the same vein as the one in libqmatrixclient.