Inviting and kicking; show rooms in Invite and Left state
Keeping rooms in Left state allows to quickly re-join but means that Left rooms may clutter space in the long run. Probably it makes sense to have some cleanup menu item for that. Also, the new code of addRoom() -> updateRoom() is more streamlined and deals with all states, thanks to a unified form for all the three signals it processes.