Allow the user to choose the account to join the room under
When `Join room...` menu item is clicked, a dialogue box pops up allowing to choose the account in cases when it can't be ascertained from the current state (namely, when there is more than one account and there is no room currently selected). `/join` will similarly connect through either the current room's account or show a request to choose the account. In any case, when an account is resolved from the current room, an additional confirmation is requested from the user, to make sure he joins using the right account. Closes #202.