KChatEdit: better history management
The changes boil down to two things:
- Do not make history entries children of KChatEdit - this does not help
to clean them up because KChatEdit::setDocument() only deletes a child
document when replacing it with another document.
- Be more diligent in saving history - editing a history entry does not
eat away both the original and the edited instance from the history
(closes #231).
A corner case remaining is when a user gets a history entry, edits it,
then moves to another history entry. In that case the edited history
entry is written over the original one so that the user could return to
editing it. However, the original history entry is gone so the user
won't be able to retrieve it even after sending the edited case.
The best probably would be to match bash behaviour that, apparently,
stores all partial edits and upon sending pushes them back to history
while making the original history entry available where it used to be
(the other partial edits still remain though).