Fill in the timeline from the bottom
This is the same technique used in Tensor and its derivatives - it's
generally more natural for chat-like applications but doesn't reasonably
support sections, which was the reason why Quaternion stayed with
TopToBottom rendering. The sections now are implemented as a separate
role in the model and an optionally shown Rectangle+Label in QML; the
"HUD" banner is a completely separate custom control now.

Most important, the whole "stick-to-bottom" workaround is completely
gone thanks to this flip - sticking the timeline to bottom now just
works.

There are still some issues from this:
- when backlogging (i.e., slowly scrolling from older messages to
newer), images and download controls lead to abrupt shifts of the
timeline in the "newer" direction. This may be a side-effect of using a
Loader with height: childrenRect.height, which effectively prevents the
engine from forecasting the delegate height. The solution is to set the
loader height explicitly, without relying on the loaded content (but
this in turn may be problematic for download controls, unless they are
actually created, which renders the Loader useless).
- the details area now pushes the whole message up, rather than
unfolding itself down. One solution to that would be to position the
details area itself above, not below, the message. This is a bit
counterintuitive (but might - or might not - turn out to be better from
the UX perspective). Another solution would be to show the details area
in the way Riot does - filling the whole timeline. This is definitely
worse UX-wise, since it enforces the area modality (and what if I want
to compare two event sources? I'm out of luck then). Yet another
solution (and the one I like most) is to make the details area a
separate popup window that can be moved away wherever the user wants,
opened several times etc. etc. The downside of this is a potential
clutter and "lost" details windows (cases when a user opened details,
switched back to the main window and forgot about the details window
behind the main one) - I have never seen this downside to cause much
inconvenience though.
Closes #217. Closes #266.