chat.qml: disable item height animation - it breaks scrolling
When a user scrolls down, new delegates are created - but since there's
animation in place, the original height of those delegates is almost 0.
Therefore scrolling causes more delegates to be created than needed.
Then animation kicks in and unfolds all those delegates pushing what's
above further to the top. Overall the effect is like "oops, I didn't
want it to scroll THAT far; and it doesn't behave the same when I scroll
up, only down".