Show the main window earlier if splash is not enabled
QSplashScreen::finish() is called even if the splash screen disabled.
That includes a processEvents() call, which in turn invokes reading
Matrix accounts and loading rooms from the cache because
MainWindow::invokeLogin() is already sent to the event loop. Loading the
cache can take seconds, during which there's nothing on the screen (if
the splash screen is suppressed). The change will remove that early
processEvents() invocation, so the cache will be loaded with the main
window already displayed.