RoomListModel: factor out ordering logic; finally (hopefully) fix crashes
AbstractRoomOrdering and classes derived from it are now used to order rooms into groups inside a model (might go to a separate pair of files in the future). replaceRoom->addOrUpdateRoom has been properly rewritten, using the latest library code to correctly track rooms lifecycle (which was the cause of crashes). Closes #402. Closes #405.