RoomListModel: Cache room indices
Instead of trying to catch up with changing room groups and positions by calculating the previous state (which proved to be terribly fragile), just store last known indices and update them along with the model when signals from Room/Connection come. The code is much simpler but consumes a bit more memory due to having to store the hashmap of rooms to persistent model indices.