Support custom chatting font
Using the same font for the timestamp, author, message items and
input box creates a unified look.

The timestamp's font is not smaller anymore. The main reason behind
this decision is that it's challenging to perfectly align the
timestamp with different sized font to the first line of the
message which could be bothering. I don't think it's worth it.

Co-Authored-By: Denis Danilov <ddanilov@users.noreply.github.com>