Timeline: hand back the focus to the input line ASAP
There is no doubt that the most clicked widget is the timeline. We
click it to open pictures, to bring the window forward, to select
text, etc. Unfortunately, these actions ended up with unfocused
input line which was my primary reason to go ape on Quaternion.

This patch introduces a strict focus policy on the timeline which
means the timeline won't keep the focus for long, it will hand back
to the input line when possible, when pressed mouse buttons are
released.

At the same time we had to implement our own TextEdit text
selection mechanism because the original didn't pass through
required mouse events. Added bonus that deselection now works for
different messages (closes #300).