Timeline: restore `Item#message` ability to shrink its height
This got broken by dbda393 which introduced circular dependency on
height: `Item#message` height was dependent on
`childrenRect.height` but it included a `MouseArea` which was
dependent on `parent.height`.

It is fixed by consolidation of three MouseAreas into one under the
top-level `Item`. This was suggested by Kitsune Ral.