TimelineItem: make ToolButtons's height fitting
Line spacing could change on the fly if the conditionally visible
buttons at the end of the line were taller than the text itself. It
could be annoying, especially for `discardButton`, because it is
visible at the beginning so each sent line's height shrinked once
which killed smooth scrolling animation. This problem was very
visible if smaller fonts were used for chatting than the default.