KChatEditPrivate::saveInput: don't clone the doc
Cloning for history the document that was just sent retains current
formatting in the entry line. As we're moving to using rich text
by default, this leads to unintended consequences: e.g., the hyperlink
for a mention at the end of the previous message gets "stuck" for
the new message (typing new characters extends this hyperlink). This
commit makes KChatEdit to create a whole new document to flush whatever
state the previous document was sent in.