Up Travis CI, next to AppVeyor
Build AppImage on bionic instead of xenial (this should tackle #562
a bit), with Qt 5.14 (to enable Markdown parsing, and it also seems
the best version for the time being). Building on xenial is thereby
abandoned (closes #710). Clang job is now used to check the oldest
supported Qt (5.9), making Travis CI one job smaller.