Redesigned shuttle dial
Instead of replacing the traditional scrollbar, the shuttle dial is now
displayed next to it. The user still can switch it off and use only
the scrollbar.
Another change along with this is in the way the shuttle dial scrolls
the timeline: instead of flicking a SmoothedAnimation instance is used.
This provides much more uniform scrolling experience and is overall
less hacky.