QML: Better sizing for room avatar
The way room avatar was sized (with sourceSize effectively bound to
the timeline size and a fragile Binding to keep image width in check)
led to unnecessary image reloads every time the Quaternion window was
resized, with topic height animation further exacerbating the problem.
The new code limits the width in a more reliable way by using
implicitWidth, and sourceSize is bound to something more fundamental -
basically, the default font height.