Room/user list: bind avatar sizes to text metrics and DPI
Instead of hard-coding avatar sizes, they are now calculated basing on
the font height and DPI as reported by QApplication. Closes #649.

P.S. Ideally these should be calculated from font metrics and DPI for
a particular widget but if we don't want to couple the model with
the view, that would mean somehow passing the same number to the model
upon creation, and updating when that changes. Maybe some other day.