ChatEdit: fix 'Reset formatting' action status tip not being displayed
I had to use this ugly way of declaring an action, because the parent
object cannot be specified in QMenu::addAction(), which led to
the status tip not being displayed.

Unfortunately, this doesn't fix the shortcut issue -- the QShortcut
still has to be in the constructor to trigger the Ctrl+M shortcut.