RoomListModel: update the way unread status is displayed
First off, the server-side unread notifications counter (bound to
the read receipt position) is used to determine whether the room is
"unread" (the name displayed in bold); previously it was the number of
unread messages since the fully read marker (which usually coincides
with the read receipt position when using Quotient pre-0.6.8 but is
different with newer library releases).

Second, the unread indicator shown after the room display name may now
have two components, shown as `[A+B]`, when the fully read marker is
behind the read receipt: A is the number of messages (ok, "notable
events" - see isEventNotable in room.cpp if you're really interested)
from the fully read marker to the (last) read receipt (counted by
the library - sic), and B is the number of messages since the read
receipt, as reported by the homeserver.

These two components are actually counted a bit differently for now,
and are only given as an estimation of how much someone needs
to scroll/read through (Quotient 0.7 and Quaternion 0.0.96 will rectify
that).

Component A may be followed by a question mark, which represents what
a plus sign previously did: an assumed number of unread messages since
the fully read marker, when the event the fully read marker points to
is not loaded. When the read marker coincides with the read receipt,
component A equals to zero and is not shown the indicator may look like
`[+B]` or `[?+B]` in that case.

Conversely, when the homeserver reports zero unread notifications
(in particular when the read receipt is at the bottom), component B
is not added, with the indicator looking like `[A]` or `[A?]`.

Finally (does anyone still read that? let me know, just curious),
there's a new debug setting that allows to visualise exactly what is
happening with the read receipt and the timeline: setting
`read_receipts` in the `[Debug]` section to `true` will add another
indicator to _all_ rooms in the room list, consisting of two numbers
in angle brackets: `<M|N>` (nothing to do with quantum mechanics,
really), where M is the number of _locally loaded_ events (not messages)
before the read receipt and N is the number of events after the read
receipt. In a corner case when an event of the read receipt is not known
M == 0 and N == the size of the locally loaded timeline; and the other
way around, read receipt at the very bottom will show as N == 0 and
M == the rest of the locally available timeline. If it happens to be
useful for non-debug cases, it will make its way from under the debug
switch (maybe in a shifted shape).