Drop --debug option
It's been doing nothing for years, and there's no clear way of making
it useful. From the looks of it, the option name is ambiguous: is it
to enable execution of Quaternion with a debugger? Extra checks? Debug
logging, as #415 hints (though one switch for debug logging is
not really enough)?

This commit drops the option and the related functions (that do
effectively nothing). Closes #415 in its current form; #787 is created
to track logging categories support in Quaternion and also introduction
of handy command-line switches with better names than --debug.