Split TimelineWidget out of ChatRoomWidget; simplify ActivityDetector
After recent changes in the way room pointers are managed (see a78ec9e0)
Quaternion started to crash on exit in a relatively reliable way. Turned
out that the discrepancy between the lifecycles of the event model and
QML view leads to QML view outliving the model and its attempts
to update bound properties during the model object destruction lead
to dereferencing what is no more an event model object. This led to
one long intended refactoring: the timeline widget is actually almost
self-contained and minimally interacts with the rest of ChatRoomWidget.
Hence TimelineWidget is now a separate thing, owning the event model
and correctly deleting it from QObject desctructor - i.e. when the QML
view does not exist anymore.

While refactoring this, it became clear that ActivityDetector has
an unnecessarily tight coupling with the rest of the classes and
moreover is created nowhere else but in main.cpp although there's no
real reason to do that (qApp static variable is available everywhere).
Making it a member of TimelineWidget simplified logistics and this class
is also much better self-contained than before.