TimelineItem.qml: view -> chatView
TimelineItem was originally intended to not depend on specifically
chatView and exposed a separate `view` property to find its actual view.
This did not materialise; these days the read marker (e.g.) code of
TimelineItem is pretty tightly coupled with chatView; and chatView is
already used instead of view in quite a few places throughout the file.
This commit accepts the fact; if a more generic TimelineItem component
will be needed, certain things will have to go to chatView.delegate
and the `view` property will be reintroduced.