systray: Notify user once between two activations of the main window
This cannot be done without leaving some indication to the user that a
notification has been issued in the past, which may anyway appear as a
lack even when notifications are issued with every new message.

This is achieved by changing the icon semi-permanently (until the next
activation of the main window), and by silently updating the tooltip
when new notifications are issued with the total number of unread
messages.

This icon change is included in the minimal notification mode, as it is
very non-intrusive.