Consolidate and refactor packaging
The whole packaging code is now gathered in a single step that has
common pre- and post-actions. This allowed to drop the dedicated
'package' option in Build matrix - 'own-quotient' is (again) reused
to perform packaging instead.

Also, instead of reconstructing the file name for each package to
download the artefact the OS name is used as the artefact name, and
a further upload just takes everything (which is always one file) from
the download directory.