/invite @natego:matrix.org
/invite @natego:matrix.org
<!-- 

This is a bug report template. By following the instructions below and
filling out the sections with your information, you will help the us to get all
the necessary data to fix your issue.

You can also preview your report before submitting it. You may remove sections
that aren't relevant to your particular case.

Text between <!-- and --​> marks will be invisible in the report.

-->

### Description

Describe here the problem that you are experiencing, or the feature you are requesting.

### Steps to reproduce

- For bugs, list the steps
- that reproduce the bug
- using hyphens as bullet points

Describe how what happens differs from what you expected.

Quaternion dumps logs to the standard output. If you can find the logs and
identify any log snippets relevant to your issue, please include
those here (please be careful to remove any personal or private data):

### Version information

<!-- IMPORTANT: please answer the following questions, to help us narrow down the problem -->

- **Quaternion version**:
- **Qt version**: <!-- for Linux systems, it's usually installed system-wide; for Flatpak/AppImage/etc.
containerised environments, it's a version used in the container or supplied with.
For other OSes, this can be skipped, as long as you provide the Quaternion version and the install method. -->
- **Install method**: <!-- package manager/Flatpak/archive downloaded (from which site?) -->
- **Platform**: <!-- Operating system and anything about your platform you think can be relevant -->