Stop using Accounts singleton
Singletons are increasingly seen as an anti-pattern that introduces
very strong coupling across the codebase. Quotient::Accounts ended up
being a prominent example of such coupling causing difficulties - e.g.,
interfering with clients making their own registries with different
account/connection lifecycle management strategies.

This commit introduces a member in MainWindow that points to Accounts;
all other calls to Accounts are replaced with calls to this member.
All classes that use Accounts for now are actually MainWindow children
and only use Accounts at initialisation; so it's quite easy to pass
the pointer to the registry as one more argument to the respective
constructors.

Eventually, MainWindow will have an AccountRegistry instance of its own
instead of using the library singleton.