Use own invokeLogin() instead of that in Quotient::AccountRegistry
AccountRegistry::invokeLogin() is deprecated as it is too opinionated
and tries to do too many things at once (collecting account settings,
creating Connection objects, and validating access tokens, to name
a few). So it is replaced with MainWindow::invokeLogin() (as it used
to be before) that, in turn, delegates the retriying loop to a newly
introduced class, ConnectionInitiator (which might be worth putting
into libQuotient further down the road).