LoginDialog: reuse saved accounts
For saved but not logged in accounts, using an initial device name that
is different from the saved one causes the homeserver to generate a new
device, which may be undesirable, especially from the E2EE perspective.
This commit makes LoginDialog to load the initial device name and
the device id for a specified account if that account is saved but not
logged in yet. The device id is now shown in UI (TODO: tuck the device
id and the homeserver behind some "Advanced" button); changing
the initial device name clears the device id (to reflect the homeserver
logic).

To make the login experience even better, two more things would be great
on top: making the user id a combo box so that the user could browse
saved account names; and adding a "Forget" button to completely remove
the saved account from the configuration.