Check for image MIME data before HTML data
Copying an image from Firefox attaches some HTML to the image.
Extracting the plain text from the HTML before the image results in
empty data in the field, with no image, triggering an assertion failure
on message send.

This is a partial fix for issue #867.  Partial, because the resulting
image does not appear in the chat window until relaunching, due to an
error opening the temp file created by this.