Make timeline scrolling less sluggish
Behavior on chatView.contentY seems to be the primary reason of
underscrolling (#711): any kind of animation in it either cancels or
"smoothens" (in case of SmoothAnimation) the previous unfinished one,
effectively dampening the reaction to subsequent wheel "clicks" coming
from touchpads and high-precision mice. This commit keeps the Behavior
disabled most of the time, allowing to mark individual self-contained
scrolling actions (sic - not response to individual input events) as
animated, e.g. when the user presses Ctrl-PageUp/PageDown or clicks
on the UI button to scroll to the last read event.

In line with that, scrollUp()/scrollDown() is only used for animated
scrolls (and one is finally implemented in terms of the other);
non-animated scrolls boil down to a change applied to contentY.

The conversion ratio between angleDelta.y and dy has also been adjusted
and bound to the actual line height.

Closes #711.