CI: Fix OpenSSL usage on Windows
Some undocumented change in the most recent Windows runner (or in a Qt
build - no idea) led to CMake getting misled and trying to use
an incomplete OpenSSL installation coming with Qt even when
tools_openssl is not chosen in aqtinstall. Given that it's always better
to use the same version of OpenSSL as that used by Qt, tools_openssl
is now actually installed on Windows.