Better ImageProvider->ThumbnailProvider
The class handles avatars in a special way now - when passed a room id
and, optionally, a user id for a member of that room, it spawns
a newly introduced AvatarResponse that can retrieve the possibly
already cached avatar from the respective Room/User object. This
required some changes in QML, with common code moved out to the special
Avatar specialisation of Image component.

What warranted a dedicated QML component is that Quotient::Avatar
returns the first (possibly empty or smaller resolution) image
immediately while requesting the right one behind the scenes; so
QML Avatar listens on the respective signals from the library and
updates the image when the new avatar arrives - which (i.e. updating) is
in turn somewhat "idiomatic" because of QTBUG-14900. QML Avatar
encapsulates these two peculiarities.