Fix a race when switching rooms at different accounts
That atomic pointer to Connection in ThumbnailProvider was never enough
to protect against the old connection being used with the new current
room, or vice versa because of a too early, or too late, request for
an avatar from QML around the moment of switching to the new room.

This commit introduces an interim null room state in
MessageEventMode::changeRoom() - this way, the old QML objects have no
chance to issue requests over the connection that already got updated;
and new QML objects only come to exist when the connection is set to
the new value.