Fix timeline items occasionally smashing into one another
After removing the animation in c074213d the problem that it was masking
manifested: because timeline items may have different height depending
on the size of the avatar (that gets loaded very much asynchronously)
it is possible that, for any given avatar on the screen, positioning of
some items in the ListView assumes the unloaded avatar height (zero)
while other items get positioned when it's already loaded and has a size
that affects the size of the item it's in. I would say that it's a bug
in Qt code for ListView (I didn't get to submit one - not yet at least)
but I can also understand the explanation that the positioning process
has to converge somewhere and restarting it when any item on the screen
changes size is probably a bit too brutal (although, again, totally not
unreasonable I think).

Anyway - the new code makes an assumption that any non-empty avatar
will take it's two-lines-of-text vertical space it is given, and simply
checks whether the avatar URL for the member is empty. It's not possible
to do it straight from QML with the current library because
Room::memberAvatarUrl() is not Q_INVOKABLE (no idea what I was thinking)
so there's one extra role in the events model that exposes whether
the sender's avatar is valid or not.

To also simplify anchoring, if the avatar is there then the "action
event" message ("user@server did something") is now simply spread across
the two lines (and it can safely spill over more lines if needed)
instead of being vertically aligned with the avatar. I think that also
removes one occasional binding loop I've seen warnings about.