ChatRoomWidget: Drop unused *drop() methods
They were used to make drag-n-drop work on the timeline area -
unfortunately the respective DropArea in the QML code was, uh, dropped
from the code accidentally, just before 0.0.95 was released. This
functionality will be reinstated in 0.0.97 in a more unified way - these
functions effectively duplicated what ChatRoomWidget was already doing
for drag-n-drop on the message input area.