.clang-format: more fixes and tweaks
The configuration of IncludeCategories so far has been copy-pasted from
libQuotient some time ago but didn't make sense as is, in the context of
Quaternion. This commit fixes the category regexes and also groups
#includes from Quotient and Qt into respective clusters, making the
whole list more compact.

Also, bring back the space after template keyword - looks like Qt Core
is sticking with it for now.

[skip ci]